Broke out auth code
